There is a hard limit per req so we need page aware polling.
This also fixes the bug where all products get accidentally
clobbered to not available then reset on the next sync.

A part of this fix was removing duplicate products as well,
by initializing the the mode based on the vendor+title
string vs the shopify product id which can change.